Add photo detail screen and navigation
Introduces PhotoDetailScreen to display photo details and updates navigation logic to allow transitioning from the home screen to the photo detail screen. MainRouteProvider is updated to register the new route.
Hero animation is used in the transition, using the thumb url as a placeholder to ensure a smooth transition to the full screen photo.